Fix Hi RAM to use 2 pins from PA not 3 (PA5-PA6) and A0-A12 to have 4*8K banks. Added GoTo option for memory navigator in clementina computer. Fixed some issues do to change of peek / poke methods to uint32. Refactor options window to allow programatically to go back to previous menu. Add tests for clementina computer. Open the coverage report after running make coverage

+++ pkg/computers/beneater/computer_test.go
@@ -342,19 +342,19 @@ func simulateKeyPress(computer *BenEaterComputer, context *common.StepContext, k
 }
 
 func validateMemoryWindow(t *testing.T, computer *BenEaterComputer, context *common.StepContext, memoryWindow *ui.MemoryWindow) {
-	assert.Equal(t, uint16(0x0000), memoryWindow.GetStartAddress(), "Window should start at address 0x0000")
+	assert.Equal(t, uint32(0x0000), memoryWindow.GetStartAddress(), "Window should start at address 0x0000")
 
 	simulateKeyPress(computer, context, tcell.KeyDown, ' ')
-	assert.Equal(t, uint16(0x0008), memoryWindow.GetStartAddress(), "Window should scroll down to address 0x0008")
+	assert.Equal(t, uint32(0x0008), memoryWindow.GetStartAddress(), "Window should scroll down to address 0x0008")
 
 	simulateKeyPress(computer, context, tcell.KeyUp, ' ')
-	assert.Equal(t, uint16(0x0000), memoryWindow.GetStartAddress(), "Window should scroll up to address 0x0000")
+	assert.Equal(t, uint32(0x0000), memoryWindow.GetStartAddress(), "Window should scroll up to address 0x0000")
 
 	simulateKeyPress(computer, context, tcell.KeyPgDn, ' ')
-	assert.Equal(t, uint16(0x00A0), memoryWindow.GetStartAddress(), "Window should scroll down to address 0x00A0")
+	assert.Equal(t, uint32(0x00A0), memoryWindow.GetStartAddress(), "Window should scroll down to address 0x00A0")
 
 	simulateKeyPress(computer, context, tcell.KeyPgUp, ' ')
-	assert.Equal(t, uint16(0x0000), memoryWindow.GetStartAddress(), "Window should scroll up to address 0x0000")
+	assert.Equal(t, uint32(0x0000), memoryWindow.GetStartAddress(), "Window should scroll up to address 0x0000")
 }
 
 func TestMenuOptions(t *testing.T) {
